Gravity-based Pulling!
The gravity wells now pull the player towards them proportional to their distance apart. The force increases exponentially until a point of no return.

Things to consider: In order to really tweak how the game feels, plenty of playtesting needs to be done. This includes player movement values as well as well pull and range. Additionally, the play field may need to be expanded or changed to facilitate a better game feel, as increasing the range of the point-of-no-return means increasing the influence of the gravity wells.